+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.4.0]
+
 ### Added
 
 - Add `sensitive` struct for redacting secret values from validation errors ([#41](https://github.com/MetaMask/superstruct/pull/41))
@@ -653,7 +655,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/MetaMask/superstruct/compare/v3.3.0...HEAD
+[Unreleased]: https://github.com/MetaMask/superstruct/compare/v3.4.0...HEAD
+[3.4.0]: https://github.com/MetaMask/superstruct/compare/v3.3.0...v3.4.0
 [3.3.0]: https://github.com/MetaMask/superstruct/compare/v3.2.1...v3.3.0
 [3.2.1]: https://github.com/MetaMask/superstruct/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/MetaMask/superstruct/compare/v3.1.0...v3.2.0